compile time dependency injection

+++ app/uk/gov/hmrc/gform/models/DateExpr.scala
@@ -65,7 +65,7 @@ object DateExpr {
 
       val dateWithOffset = LocalDate
         .parse(dateExprToString(dateExpr), dateFormatter)
-        .plusDays(nonZero)
+        .plusDays(nonZero.toLong)
 
       convertToDateExpr(dateWithOffset)
   }
